mb705 reset
sh_stm: Use mb705 reset signal for mb680 MII

When used standalone, the mb680 uses a PIO for MII reset, which is what
the previous code assumed. However when used with an mb705 the reset comes
from an EPLD register on the mb705. So provide a generic interface for
the mb705 reset function, and wire it up to the MII.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>