sh_stm: STMicro Cut3 7105 REQ/GNT PCI changes
Cut3 of the the 7105 and cut2 of the 7141 introduced a change as
workaround for a problem sharing the PCI and NAND. The workaround is to
internally connect what was REQ/GNT0 to REQ/GNT3. This drops the number
of external masters to 3 instead of 4, but this is not expected to be a
problem for the vast majority of potential applications. This change is
internal to the chip, the pinout is the same.

This code sets up the EMI arbiter correctly in the case of cut3. This is
done by probing the chip type and no changes to the BSP are needed. This
means that PCI cards that use bus mastering (virtually everything)
should now work correctly on cut3, whereas before they would never have
been granted accesss to the bus.

Signed-off-by: David McKay <david.mckay@st.com>