sh_stm: Fix STmicro PCI Reset PIO func
Some boards like the PDK 7105 allow software control of the PCI reset as
well as generating a power on reset to the PCI. This is done via a diode
to prevent the software PCI reset resetting the board. Unfortunately if
the software PIO reset actively drives the line high, it prevents the
reset circuit from pulling the line low, meaning that reset doesn't work
properly.

The solution is to either drive the PIO open drain or change the PIO to
an input. The reset line will be pulled high by the pullup resistor.
Since the existing code uses the gpio interface, we change the PIO to an
input as the open drain mode is not available via the gpio interface.

Thanks to Michael Schenk for spotting this.

Signed-off-by: David McKay <david.mckay@st.com>