FAT: Add CONFIG_VFAT_NO_CREATE_WITH_LONGNAMES option
When this option is enabled this will refuse to create new files with
long names. Accessing existing files with long names will continue to
work.

File names to be created must conform to the 8.3 format.  Mixed case is
not allowed in either the prefix or the suffix.

See http://kernel.org/pub/linux/kernel/people/tridge/VFAT for more
details.

Signed-off-by: Andrew Tridgell <tridge@samba.org>
Signed-off-by: Dave Kleikamp <shaggy@linux.vnet.ibm.com>
Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>