LKM: Fast loader - Add ksymhash tool
Add the ksymhash tool to the kbuild. It is used to compute
the hash values of the exported symbols (vmlinux and modules) and
the hash values of the undefined symbols (only for modules).
The later will be include into the __ksymtable as part of the
struct kernel_symbol and the latter into an array placed in the
ELF section .undef.hash
So there will be no need for computing hash values at runtime.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>