LKM: Fast loader - Extend modpost to use elf hash feature
Extend the modpost tool to use the new support functions for creating
the elf hash table related to each __ksymtab and the .undef.hash section.
The .undef.hash section at this stage is just a place holder because the
final link of the module is not yet actually done. The section will be
properly filled with the hash values in a extra kbuild stage by ksymhash tool.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>