stm: irq: Updated ILC device driver
This patch updates the ILC device driver to be able to manage
more than one ilc device.
This is required in the 7141 platform where linux uses the eSTB ILC
to manage standard interrupt but it _has_ to use mandatory
the comms.ilc device to really manage the wakeup_out signal.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>