sh: Kernel threads should not inherit FPU flags from parent
Prior to this patch kernel threads would inherit the FPU usage flags
from their creating process. This was not a problem because nothing
ever looked at these flags for a kernel thread. However commit
33806d4f8804a5ec8a6718b08dfdf5c57b7e29eb "sh: add sleazy FPU
optimization" changed this, because __switch_to() now looks at the
FPU flags when determining whether to load the FPU regs before
executing the thread.

This patch clears the FPU flags when creating a kernel thread.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>