sh: Move the alignment exceptions flags
Commit f775a6d40e4812ce9f97af7b589cb8e26c3275e3 "sh: implement
PR_[GS]ET_UNALIGN prctls." introduced three new task flags (TIF_...)
which are used to record the misaligned handling behaviour requested
by the user through the prctl command.

However the flags were positioned in the lower byte of the flags word,
which means that they are checked when returning to user mode against
the work flags (_TIF_WORK_MASK and _TIF_ALLWORK_MASK), and they
shouldn't be. We could just change the masks, but as later kernel
versions use these bit positions for things which do have to be
included in the work masks, and the assembler code can be more
efficient if the work masks are 8 bits or less, move the misaligned
flags to the top half of the word.

Also drop the NOFIX flag as it was never used.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>