net: added missing fields to have trans_start
This patch adds some missing fields into the netdev_queue
structure in order to update the trans_start into the layer.
This patch has been added to test the stmmac driver without
the dev->trans_start in the xmit.
This patch must be removed marging the new trees because they
already support trans_start in the layer.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>