stmmac: tidy-up the vlan support.
Currently the driver is only able to detect if the incoming
or the transmitted frame is vlan tagged.
In the past, an incomplete code for filtering rx vlan frames
was added but never tested due to HW limits.
This code is now removed; it will be re-written from scratch
as soon as we could actually test the filtering.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>