stmmac: reviewed locking, transmission process and funct args.
This patch essentially reviewed the locking code for
the transmission process. The stmmac runs in softirq and
the tx resources are claimed though poll (NAPI);
so unnecessary spin_lock(s) have been removed.
Main work is stolen from the tg3 driver.
Moreover, there was a problem while enabling tso
support due to a wrong locking protection.
The patch also reviews the logic behind the tx
clean process. Now, during polling we directly invokes the
stmmac_tx function that reclaims and cleans the tx ring.

Finally, this patch  reviews the arguments passed to many
functions. Indeed, it's better to pass the private structure
instead of net_dev.

Note that after applying this patch, I've also noticed
an improvement of the performances due to the wrong
logic adopted in NAPI when invoke the tx clean function.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>