stm_pio: New stm_pio10 driver
STx7105 introduced a new style of PIO block, with 10 normal STM style
PIO ports, but with all the interrupts OR'ed together and a new register
indicating which PIO bank is currently asserting its interrupt.

Add support for this hardware, with a new device: stm_pio10. This makes
use of the stpio driver for most functionality, with a new top level
interrupt handler to decode the new status register.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>
Acked-by: Pawel Moll <pawel.moll@st.com>