libata: Don't try and select slave if its not possible
ata_sff_postreset() tries to select both drives to check whether a
drive exists. If however we have already determined that the controller
doesn't support a slave drive, then the slave drive class is
effectivly undefined, and reads as 0 which is ATA_DEV_UNKNOWN.

Change the code to only try and select the slave drive if it is known the
controller supports it.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>