LKM: Fast loader - fix dependency for .tmp_vmlinux
The temporary vmlinux used to generated the hash table does not actually
depend on the linker script vmlinux.lds so it can be safely removed.
Indeed the vmlinux-modpost rule does not filter it out any more; this change
exploited a problem with latest binutils 2.19 (used in STLinux-2.4) that did
not occurr with previous version 2.18 (used in STLinux-2.3).

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>