kbuild: Allow to use depmod from a non default path
This patch allow to specify the directory where to find the
depmod tool other than being forced to use the module-init-tools
installed on the host. The reason for this is explained below.
When doing cross-compilation, it not guaranteed that the
module-init-tools installed on the host (and used during kernel build)
are exactly the same and compatible with those used on the target at runtime.
For example, recent changes in how depmod writes modules.dep using
root-less path, make older modprobe unusable on the target.
User in ths case could install in a different path the module-init-tools
compatible with the version used on the target and specify the path on the
command line.

Signed-off-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>