sh_stm: fix SSC sysconfig claim on STi7111
Earlier versions of the 7111 datasheet erronously claimed that the ssc
used bits 9-11 in sysconfig 5 were used to control the SSC. Actually it
is bits 10-12. PCI wants to control bit 9 to get the PCI INT0 interrupt,
so these have to be changed. If you want to use INT1-3, you have to not
configure SSC1.

There were also some sysconfig bits intended to control ASC/SCIF
multiplexing, but the final device shipped without any SCIFs, so these
can be removed.

Signed-off-by: David McKay <david.mckay@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>