stm_fdma: Add additional checks to the firmware loading
Currently we implicitly trust the addresses passed in the FDMA
firmware. Add knowledge to the FDMA driver of where the RAM regions
are, and check that the firmware is only loaded into those memory
regions.

Also check that we really are loading SLIM code, and it is really is
for the FDMA.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>