sh_tmu: add new APIs to use a TMU channel as generic timer
A TMU channel not used as clock source or for clock event (e.g. TMU2)
can work as generic timer and can be used by another device (that
want to have the TMU interrupt to invoke own handler).
An example is the stmmac Ethernet device driver. It can use the sh_tmu
timer for mitigating the number of interrupts.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>