gpio_stm: Intergrate gpio & PIO10 drivers
This patch changes the way that the stand-alone PIO10 interrupts
are being handled - instead of "stm-pio10" device & driver, handling
both PIO ports and interrupts, there is a "stm-gpio-irqmux" driver,
which handles the interrupt muxer only, and PIO ports are represented
by separate standard "stm-pio" devices.

This also prepares the driver for the "pad manager" rework.

Signed-off-by: Pawel Moll <pawel.moll@st.com>