stm: Pad manager rework
This patch reworks the pad manager "ecosystem".

Most notable changes are:
- idea of "labels" has been totally abandoned, as was proofed
  to be pointless and annoying
- pad manager & GPIO drivers are aware of each other existence
  and relations (thus a private API for GPIO driver in pad manager)
- an idea of "PIO function" has been introduced - it is just
  a logical representation of different PIO uses (widely known
  as "alternative functions") and totally SOC specific (transparent
  to pad manager)
- a SOC-specific callback to configure given PIO has appeared,
  which
- pad manager must be initialized by the SOC-specific code now,
  what also defined PIO configuration function mentioned above

Note that this patch updates all drivers using pad manager,
but _DOES NOT_ update SOC setups (which, due to size of changes,
are going to be split into separate patches), so the code will
not compile correctly at this point.

Signed-off-by: Pawel Moll <pawel.moll@st.com>