stm-asc: Only check for parity errors when enabled
Checking the ASC documentation, it turns out that the parity error
bit on recived characters is only defined when parity has been enabled:

  If the MODE field selects an 8-bit frame then this bit is undefined.
  Software should ignore this bit when reading 8-bit frames

So add a flag to allow the recive code to determine whether it should
check.

This was found through bug report:
https://bugzilla.stlinux.com/show_bug.cgi?id=8174
although this is a different implementation of the fix.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>