stm: cpufreq: Update cpufreq support for the STx5197/5206/7105/7111/7141
Update the cpufreq support for the 5197, 7105, 7111 and 7141,
and add support for the 5206. This follows the previous update
to the STM specific cpufreq code code, using the new interface
between the generic and SoC specific code.

All these chips have a complete Clock Low Lever API support
for this reason the CPUFreq system just calls the clk_set_rate(...)
on the right clock according to the chip.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>