stm: cpufreq: Update cpufreq support for the STx7200 c2
Update the cpufreq support for the 7200. This follows the previous update
to the STM specific cpufreq code code, using the new interface
between the generic and SoC specific code.

Unfortuntaly this SoC isn't currently supported fully by the clk framework
so this still manipulates the clockgen hardware directly.

Currently the cpufreq support on 7200 c1 isn't available.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>