stmmac: rework normal and enhanced descriptors usage
Currently the driver assumes that the mac can only use the
normal descriptor structure and the gmac wants the
enhanced structures.
With patch removes the descriptor code from the dma part
and adds two new file for handling the normal and enhanced
descriptors.
So this patch helps mac and gmac to use either normal or
enhanced descriptors structures.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>