stmmac: new descriptor field for the driver's platform
The new enh_desc is used for selecting the enhanced descriptors
structure. There are several scenarios; some chips (mac10/100
or gmac) want to use the enhanced descriptor; some other the
vice-versa. On ST platform the MAC10/100 uses the normal structure
and the GMAC uses the enhanced one.
So it can be useful to get this information from the platform.
This could also be decided at run-time looking at the chip's ID number.
hmm ... but it could happen that chips with the same ID want to use different
descriptor structure.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>