stm_nand_flex: Workaround for WAIT_RBN bug, GNBvd77579
Due to a bug in the FLEX NAND Controller (GNBvd77579), clearing the 'WAIT_RBN'
config bit is ignored.  As a result, any write to the FLEX command register will
cause the controller to wait until the NAND device becomes ready.  The CPU, bus,
and EMI will all be stalled during this period.

The default nand_wait() function, called after each NAND operation, starts by
writing a NAND_CMD_STATUS command.  This will stall until the operation has
completed, which may take several milliseconds in the case of an erase
operation.

This patch overrides the default nand_wait() function, and waits until RBn is
asserted before issuing the NAND_CMD_STATUS command.  If the RBn signal is
unavailable, then we revert to the default behaviour, and accept the stall.

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>