stm_nand_flex: Fix boot-mode ECC support
Support for STM boot-mode ECC relied on MTD features that were no longer present
in the 2.6.32 kernel.  This patch updates the way in which boot-mode ECC is
supported and simplifies the registration of MTD partitions.  Scanning for
bad-blocks has also been updated to correctly identify boot-mode ECC data,
removing the need to use FLASH-resident BBTs (although the use of BBTs remains
strongly recommended).

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>