stm: pms: simplified sysfs interface
This patch simplifies the sysfs interface.
Now the pms will use less command to be programmed via 'control' sysfs file.
It was removed the
 - state ...
 - device ...
 - clock ...
 - cpu ...
command.
Now the only supported command are the 'constraint' command.
The pms code will check if the 'state' and the 'object' (clk/device/cpu) isn't
still registered and if so it creates the new object on the fly.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>