sh_stm: add mmc configuration for the 7106
Tested on HDK7106 cut 1.0 plus MMC daughter board V 1.0.

Notes:

MMC is supposed to be configured as ALT_OUT, not ALT_BIDIR.
GNBvd78840 MMC/SPI interface is working not ok when PIO is set in BD
This is necessary for the 7106 cut 1.0.

There is and error on the 7106-HDK MMC daughter board V 1.0 schematics.
The STMPS2151STR Pin 4 should be 'EN' and the Pin 5 should be
'Power_IN'. These are reverse.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro at st.com>