sh_stm: temporary fix for the phy clk on the 5206/89
This Kernel network doesn't come-up on the 5206/89 when use the
delivered targetpacks (e.g. the latest stmc-1.5.0.0-45).
The problem is due to the Ethernet clock that's broken
and not good to work in MII mode (where 25MHz are required).

Waiting that it will be fixed this patch can actually help
to have the network working on this board.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>