spi_stm: workaround for RIEN interrupt bug
spi_stm: workaround for RIEN interrupt bug

Major rework of the SSC-SPI driver due to misbehaving RIEN/TIEN interrupts.  As
a workaround, we use the TEEN ("Transmit Error") interrupt as a notification of
when the TX shift register becomes empty.

The patch also adds a special case for even-length, 8 bits-per-word transfers,
which are now implemented as half-length 16 bits-per-word transfers, thereby
halving the number of interrupts generated.

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>