sh: Improve PMB mapping flexability
Currently pmb_calc() can fail to calculate a mapping even when there
is a suitable hole in the virtual address map because it has chosen
PMB entries which are too large and so have a large alignment constraint.
For an example see STLinux bugzilla 7488.

This patch should solve this particular problem by retrying the mapping
calculation with restricted PMB sizes whcih will reduce the alignment
requirement.

Also add several comments to try and make the algorithm a little
clearer - written while trying to understand the code again after a long
break!

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>