mtd-nand: Add debugfs attribute to allow erasing of blocks marked bad
mtd-nand: Add debugfs attribute to allow erasing of blocks marked bad

This patch adds a debugfs attribute to the MTD NAND support to allow erasing of
blocks marked bad, including those blocks reserved for FLASH-resident bad-block
tables.

Allowing bad blocks to be erased is normally a bad idea.  However, during
development, it may be necessary to "scrub" the NAND device (e.g. switching from
boot-mode ECC back to linux ECC, or recovering from a corrupted bad-block
table).

Note, this option should be used with caution as it may result in the loss of
factory-set bad-block markers.  It is advisable to record the existing set of
bad-blocks prior to scrubbing the NAND device

Signed-off-by: Angus Clark <angus.clark@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>