sh: Improve PMB mapping use of virtual memory space
Prior to this change, pmb_calc() would try and use the maximum size
PMB entries possible, assuming that PMBs are a scarce resource and
mustn't be wasted. However this can lead to virtual address space being
wasted, which in some cases can be a larger problem.

See STLinux bugzilla 2833 for example.

So this change introduces a heuristic: if increasing the mapping size
would result in a mapping where half or more of the coverage is being
wasted, don't try and increase the mapping size.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>