sh: Safeguard low level PMB manipulations
As setting a PMB entry involves a write to both the address and data,
arrays, ensure that an inconsistent state cannot exist between the two
writes. First clear the valid bit, then write to the address
array with the valid bit clear, and finally write to the address
array again, this time with the valid bit set.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>