sh: Add option to tile P1/P2 region with 64M PMB entries
The SH architecture behaviour that P1/P2 accesses which miss the PMB
cause a reset can be very annoying. So this change adds an option
to tile the entire P1/P2 region with 64M PMB entries, which are always
valid, and so a miss can never occur.

This then allows other debugging techniques such as data watchpoints to
be used to try and catch erroneous reads or writes.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>