stmmac: fix timer opt when use dual mac config
the driver set the tmrate to zero when the timer
initialisation failed. This generates problems with
the dual GMAC configuration.
With this patch, enabling both the dual gmac and
the timer optimisation, the first interface opened
will use the tmu channel 2, the second ethX won't
be able to use the timer but will continue to work
without mitigating the interrupts via timers.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>