stm: clk: Fix potential array overflow issue
The four ratio arrays used in the code are indexed directly by the
value read from the hardware register. However there are some undefined
values in the hardware whcih had no valid index in the ratio array,
which could have gone undetected. So fix the size of the array so all
index values are now valid, and add a BUG_ON if an invalid field is used.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>