stm: suspend: Record and restore comms_clk rate
Having introduced clkdev, and switched the comms_clk to an alias of the
real clock, the technique previously used in the suspend code of reading
the comms_clk frequency from its parent will no longer work. So record
the comms_clk rate prior to suspending, and restore it when resuming.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>