stm: Add the stm_device wrapper
This wrapper is should assist device drivers for STM specific peripherals
to hide the SoC to SoC differences. This process was started with the
pad manager and the stm_device layer adds capabilities for power
management.

For most device drivers stm_device can be used as a drop in replacement
for the current pad manager, or it can be used alongside the pad
manager, just using its power management capabilities.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>