stm: pm: Add early pm notify callbacks
This patch adds a set of callbacks around the core suspend code.
These callbacks allow to the customer an early check of the wakeup
reason, and if its erroneous can put the system to sleep again.

Typically it's used to allow early IR sequence decoding
to avoid a complete system resume.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>