stm: pm: stx7108: Improve LMI suspend support on 7108c2
This patch improves the suspend on stx7108_c2.
Now the system also turns-off the DDR_Phy.
The only difference between c1 and c2 now is in the
DDR_Phy.ODT management.
It seems the DDR_Phy.ODT has some issue in the cut 2 also
related to the setup done in the target_pack (investigation on-going).

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>