stm_fdma: Added PM support
This patch adds the PM support to the stm-fdma device.

In case of suspend-to-RAM the driver:
 - stops the channels on suspend;
 - enables the channels on resume.

In case of hibernation the driver:
 - stops the channels on freeze
 - initializes all the fmda on restore (reprogramming the firmware)

When restore from hibernation a new request_firmware fails because
the user space is not available and it is likely some devices already
claim the fmda-channels.
To fix that, a new private structure (fdma_segment_pm) has been added.
It is used for storing the firmware (text and data).
When resume from hibernation the firmware in the fdma_segment_pm is
restored again.

Hibernation was tested on stx7108c2 programming the fdma in
freerunning mode and moving data mem-to-mem.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Reviwed-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>