fdma: modify firmware name
This patch to modify the firmware name loaded by request_firmware.
It will be now called fdma_<cpu_subtype>_<X>.elf, where cpu_subtype will be
the cpu_subtype name, as kernel is able to categorize it (STx7105, STx7109,..)
and X will be the fdma number (zero if only one fdma is present).
In this way we align the naming convention with the one used in the new
fdma firmware package, distributed soon in the STLinux distro.

Signed-off-by: Giuseppe Condorelli <giuseppe.condorelli@st.com>
Reviewed-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>