stm: clocks: stx7100: Make clockgenA_clk_enable always succeed
This patch changes clockgenA_clk_enable to always return success,
even for those clocks which cannot be disabled.
This is because some critical clocks which can not
be touched will still be enabled though the clock framework,
and without this the enable will fail.

Note this makes the 7100 consistent with the other SoCs.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>