stm_pm: Fixed WoL support on stx5206
This patch fixes the WOL support on stx5206.
On some board the eth_phy has no external oscillator
and it uses a clock generated by the chip.
This clock has to remain enabled during standby to
support a working WoL.
Required to anticipate the LMI_pad reset to have WoL working.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Reviewed-by: Angelo Castello <angelo.castello@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>