stm: clks: Remove postcore_init initialisation
In the chip 7105 and 7108 (and only these chips) the clocks were
registered in 2 steps (as system-core clock in the 'plat_clk_init()'
and other clocks postponed to a 'postcore_initcall(postcore_clk_init)'...)

But to register correctly all the clock aliases, all the clocks must be
registered first.

So register all the clocks (core clocks and not core clocks) in a single
function.

Moreover in this manner all the drivers/stm/clocks/clock-stx****.c
files now follow the same design.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Reviewed-by: Giuseppe Condorelli <giuseppe.condorelli@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>