stm_pad: rework the Eth PAD, retime conf and support the RGMII mode.
This patch adds the RGMII mode tested on the HDK7108 (v1/v2)
with the STx7108 cut 2.

Below the summary of the HW changes needed on these boards:

To use the GMII mode (GTX CLK) on the HDK7108V1:
 --------------------------------------------------------
        RP18        RP42        RP14        Rp17
 --------------------------------------------------------
        GMII        NC         51R        NC        51R
        MII        NC         NC        51R        51R

  On the HDK7108V1/2: remove R31 and place it at R39.

  RGMII more requires the following HW change (on both
  HDK7108V1 and HDK7108V2): remove R29 and place it at R37

The RGMII mode is similar to the GMII (GTX) except the data
buses are reduced down to 4 bits and the 2 error signals are removed.
The data rate is maintained by using both edges of the clock.

To support this mode a new retiming configuration has been added.
In order to avoid other hacks and to make clear the whole 7108
Ethernet configuration, the PIO and retiming structures have been
totally reviewed. Now the retiming configuration is passed as parameter
in the stm_pad_config structure.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Reviwed-by: Francesco Virlinzi <francesco.virlinzi@st.com>