rtc_stm_lpc: Fix wrong LPC.LPT date on boot
This patch fixes an issue in the date management.

The LPC.LPT is able to manage date.year > 2038 but currently
Linux can not manage this date.

On some HDK-7108 boards during power-on
the LPC.LPT has date.year > 2038 and due to the kernel constraints
this date is converted to a date.year < 1970.

With this patch during the boot the LPC.LPT is checked and in case
of date.year > 2038 the LPC is reset to Jan 1st 2000.

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>
Reviewed-by: Carmelo Amoroso <carmelo.amoroso@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>