stm: Update MiPHY infrastructure and support dual SATA on 7108c2
Add support for the dual SATA hardware on 7108c2. This involves the
introduction of MiPHY interface infrastructure, and move the existing
TAP interface so that it sits behind this interface. Add the new PCIe
microport interface to the MiPHY as well, and switch 7108 to use the
TAP interface on cut 1 and microport on cut 2.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>