sh_stm: Move ILC3 interrupt registration ahead of GPIO initialisation
The initialisation of ILC3 interrupt chip and handler used to
occur in a function called early in the kernel initialisation
sequence from plat_irq_setup(), called from init_IRQ().

Subsequently the commit
  617d6e9c24936a84e6fd0b1746c6ea2b532dad18
  stm: irq: Updated ILC device driver
moved the initialisation the the regular "ilc3" platform device
probe function. Both the driver and the device were registered at
arch initcall level.

Meanwhile, stm-gpio device initialisation attempts to set up a second
interrupt chip and handler for the PIO interrupts. This occurs in
its probe function. Both the driver and the device are registered at
postcore initcall level.

This means that the ILC3 interrupts have not been initialised at
the point the stm-gpio driver tries to register its own interrupt
on top of them. This results in the error message:
  Trying to install chained interrupt handler for IRQxxx

This patch changes all the ILC3 initcalls so that they occur at core
initcall level, meaning that it is before the GPIO initialisation.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>