stm_pad: add a dummy function to help on debugging retime conf for the 7108
This patch adds a dummy function to help on debugging the retime and
gpio setting. This actually is useful while testing all the ethernet
modes (mii/rmii/gmii/rgmii etc.).

Maybe, thise could become more generic and could be moved to the
pad mng.

See bug: 11107

Turn-on the DEBUG_RETIME_CONF for enabling this support.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>