stm_mb903: Support MoCA device as a fixed PHY
The STvMoCA device in the STx7108M doesn't support enough of the
IEEE 802.3 SMI register set to allow the generic PHY device code
to probe and configure the device. On the other hand it doesn't
need to, because the device is effectivly fixed function, from the
point of view of the Ethernet device. So set up the fixed PHY
infrastructure so that the stmmac binds to a fixed PHY.

Signed-off-by: Stuart Menefy <stuart.menefy@st.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>