stm_mali: Integrate Mali drivers (r2p0-05rel0) into STLinux.
This is the kernel driver for the Mali GPU, released by ARM under the GPL.
The vast majority of this code was written by and is copyright ARM Limited.

Minor modifications have been made by STMicroelectronics to the build
system and configuration. With these changes the Mali drivers can now be
built into kernel or as a module by selecting CONFIG_STM_MALI under
"STM specific devices" from menuconfig.

Patch is tested on HDK7108 board.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>