stm_mali:Fixed Dereferencing freed vma_area_struct pointer in vma_close.
As vma_area_struct in descriptor and vma in arguments to vma_close can
be different. With this caveat, dereferencing vma to get tracker can
fault, as kernel MIGHT have freed the vma_area_struct in the
descriptor. Now vma_usage_tracker can be obtained from mappingInfo,
which will be valid at the point where its being freed.

This patch adds vma_usage_tracker pointer to mappingInfo struct as vma
pointer stored in the descriptor MIGHT be already freed by the kernel.

This bug can be easily reproduced with a multithreaded app doing 3D in
one thread and forking in other, after a while the kernel crashes.

Orignally this issues was reported in Bugzilla#10801.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@st.com>
Signed-off-by: Stuart Menefy <stuart.menefy@st.com>