stm_ssc_i2c: Set Slave Address to support general call
This patch sets the I2C slave address register for all
the SSCs to support general call addressing.

In case of general call (address equal zero) the SSC writes
the address on the bus but as this matches the slave address
register (which is initially zero) slave side of the SSC detects
the zero address and switches into slave mode like a 'standard
addressed as slave'.

The slave address register is initialized to 0x7f which is
declared in the I2C spec as 'reserved for future purpose'
therefore it's generally available.

Reference: Bug 11690

Signed-off-by: Francesco Virlinzi <francesco.virlinzi@st.com>