stm_pad: add pad mmc support to Freeman platforms
This patch adds the MMC/SD support to Freeman.

The  FLi7510 (Freeman-Lite) Development Board (FLDB_GPD202)
and the FLi7540 (Freeman-Utra) Development Board (FUDB_GPD202)
have been used for doing all the tests (card as mass storage
device).

Note: on the Freeman-Lite to use the MMC/SD card on the external
CNG6 connector, a ribbon cable must be used to connect CNG4 and CNG3.

Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Reviewed-by: Carmelo Amoroso <carmelo.amoroso@st.com>