sound_stm: Simplify SOC specific glue drivers
This patch is the second in a series designed to incrementally add working ALSA
support to the SH linux kernel. The SOC specific glue drivers no longer register
a sound device to enable and disable the audio outputs. Instead the glue drivers
enable the audio outputs on probe and disable the audio outputs on exit.

Signed-off-by: Pawel Moll <pawel.moll@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>