sound_stm: Simplify the SOC audio subsystem core driver
This patch is the fifth in a series designed to incrementally add working ALSA
support to the SH kernel. It will combine the functionality of the SOC audio
subsystem core driver with common routines and simplify the process of
registering the SOC glue driver with the SOC audio subsystem core driver.

Signed-off-by: Pawel Moll <pawel.moll@st.com>
Signed-off-by: John Boddie <john.boddie@st.com>